refac: fraud NFT order logic
